Split large taskengine test files for faster CI/CD
- Split engine_test.go (1282 lines) into 4 focused files
- Split vm_test.go (1431 lines) into 5 focused files
- Split run_node_immediately_test.go (1532 lines) into 4 focused files
- Maintained all existing test coverage and isolation
- Enables parallel test execution to reduce CI time from 8m13s

Files created:
- engine_crud_test.go (task creation, validation, listing)
- engine_execution_test.go (execution, triggers, status)
- engine_trigger_output_test.go (trigger output processing)
- vm_compilation_test.go (VM compilation and basics)
- vm_branch_execution_test.go (branch logic)
- vm_event_processing_test.go (event triggers)
- vm_preprocessing_test.go (text preprocessing)
- run_node_integration_test.go (integration scenarios)
- run_node_templates_test.go (template processing)
- run_node_triggers_test.go (trigger execution)
